times unix

+++ fze-commons/src/main/java/net/fze/extras/report/ReportParses.kt
@@ -6,7 +6,7 @@ class ReportParses {
     companion object {
         /** parse time range like [2020-05-06T16:00:00.000Z, 2020-05-08T16:00:00.000Z] */
         @JvmStatic
-        fun parseTimeRange(s: String): Array<Int> {
+        fun parseTimeRange(s: String): Array<Long> {
             if (s.isEmpty()) return arrayOf()
             var src = s
             if (src[0] == '[') {
