+++ fze-commons/src/main/java/net/fze/util/concurrent/TokenBucket.kt
@@ -5,7 +5,7 @@ import net.fze.util.Times
 // 令牌桶算法, 参见: https://github.com/ixre/gof/blob/master/util/concurrent/token_bucket.go
 class TokenBucket {
     // 时间
-    private var timestamp: Int = 0
+    private var timestamp: Long = 0
 
     // 桶的容量
     var capacity: Long = 0
